requirements.txt
flask==2.0.1
werkzeug==2.0.1
pillow==10.4.0
paramiko==2.7.2
python-dotenv==0.19.0
numpy==1.23.5
opencv-python==4.9.0.80
scikit-image==0.19.3
allure-python-commons==2.13.2
six==1.16.0 